refactor: 영화 목록 및 항목 구성 요소 폴더 분리
