+++ src/components/movie/movieList/movieList.ts
@@ -1,6 +1,6 @@
-import movieItem from "./movieItem";
-import { createElementWithAttributes } from "../utils/createElementWithAttributes";
-import { Movie } from "./types";
+import movieItem from "../movieItem/movieItem";
+import { createElementWithAttributes } from "../../utils/createElementWithAttributes";
+import { Movie } from "../../../domain/types";
 
 const movieList = (movies: Movie[]) => {
   const $movieList = createElementWithAttributes({
